requirements.txt
adrf==0.1.7
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-property==0.2.2
certifi==2024.8.30
cloudinary==1.41.0
distro==1.9.0
Django==5.1
djangorestframework==3.15.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
jiter==0.5.0
openai==1.43.0
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
